+++ ggml/src/ggml-openvino/ggml-openvino-weight-buffer-release.h
@@ -2,10 +2,14 @@
 
 #include <cstddef>
 
-// Host weight-buffer release (GGML_OPENVINO_RELEASE_WEIGHTS, GPU only).
-// register: record a host weight buffer (idempotent per data pointer).
-// release:  madvise(MADV_DONTNEED) all registered buffers, dropping their RSS.
-// released: true once release has run (used to fail-fast on post-release recompile).
+/// \brief Record a host weight buffer for later release when GGML_OPENVINO_RELEASE_WEIGHTS is enabled.
+/// \param data Base address of the host weight buffer.
+/// \param size Size in bytes of the host weight buffer.
 void ggml_openvino_register_weight_buffer(void * data, size_t size);
+
+/// \brief Release registered host weight-buffer pages with madvise(MADV_DONTNEED).
 void ggml_openvino_release_weight_buffers();
+
+/// \brief Return whether registered host weight buffers have already been released.
+/// \return True after ggml_openvino_release_weight_buffers has run.
 bool ggml_openvino_weight_buffers_released();
